fix init param idempotency
